+++ feed/index.html
@@ -13,7 +13,7 @@
     <link rel="icon" type="image/png" sizes="16x16" href="/images/favicon-16x16.png" />
     <link rel="manifest" href="/site.webmanifest" />
     <link rel="stylesheet" href="../dist/css/index.css" />
-    <script type="module" src="/js/feed.js" defer></script>
+    <script type="module" src="/js/feed.mjs" defer></script>
   </head>
   <body>
     <header class="bg-white p-4 sticky-top">
