+++ index.html
@@ -13,7 +13,7 @@
     <link rel="icon" type="image/png" sizes="16x16" href="/images/favicon-16x16.png" />
     <link rel="manifest" href="/site.webmanifest" />
     <link rel="stylesheet" href="dist/css/index.css" />
-    <script type="module" src="/js/index.js" defer></script>
+    <script type="module" src="/js/index.mjs" defer></script>
   </head>
   <body>
     <main>
