+++ js/api/auth/register.mjs
@@ -1,5 +1,5 @@
-import {API_AUTH, API_BASE, API_REGISTER} from "../constants.js";
-import { authFetch } from "../fetch.js";
+import {API_AUTH, API_BASE, API_REGISTER} from "../constants.mjs";
+import { authFetch } from "../fetch.mjs";
 
 export async function register(name, email, password) {
     const response = await authFetch(API_BASE + API_AUTH + API_REGISTER, {
