+++ js/api/fetch.mjs
@@ -1,4 +1,4 @@
-import { headers } from "./headers.js";
+import { headers } from "./headers.mjs";
 
 export async function authFetch(url, options = {}) {
     return fetch(url, {
