+++ js/api/posts/get.mjs
@@ -1,5 +1,5 @@
-import {API_BASE, API_POSTS} from "../constants.js";
-import {authFetch} from "../fetch.js";
+import {API_BASE, API_POSTS} from "../constants.mjs";
+import {authFetch} from "../fetch.mjs";
 
 export async function getPosts() {
     const response = await authFetch(API_BASE + API_POSTS);
