--- js/index.js
@@ -1,3 +1,0 @@
-import { homePage } from "./routes/home.js";
-
-homePage()
