+++ js/index.mjs
@@ -1,0 +1,3 @@
+import { homePage } from "./routes/home.mjs";
+
+homePage()
