+++ js/routes/home.mjs
@@ -1,4 +1,4 @@
-import {setAuthListener} from "../ui/listeners/auth.js"
+import {setAuthListener} from "../ui/listeners/auth.mjs"
 
 export async function homePage() {
     setAuthListener()
