+++ js/ui/listeners/auth.mjs
@@ -1,4 +1,4 @@
-import {onAuth} from "../events/onAuth.js";
+import {onAuth} from "../events/onAuth.mjs";
 
 export function setAuthListener() {
     document.getElementById("auth-form").addEventListener("submit", onAuth);
